[Refactor] #10 - NavGraph 수정
Navhost 앱 플로우 버그 수정
